Knit a set of tutorials

We define "testing" a tutorial as (successfully) running render() on it. This function renders all the tutorials provided in tutorial_paths. There is no check to see if the rendered file looks OK. If a tutorial fails to render, then an error will be generated which will propagate to the caller.

Usage

knit_tutorials(tutorial_paths)

Arguments

tutorial_paths

Character vector of the paths to the tutorials to be knitted.

Value

No return value, called for side effects.

Examples

  knit_tutorials(tutorial_paths = return_tutorial_paths("tutorial.helpers"))
#> Rendering: /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/tutorial.helpers/tutorials/getting-started/tutorial.Rmd
#> 
#> 
#> processing file: tutorial.Rmd
#> 1/71                     
#> 2/71 [setup]             
#> 3/71                     
#> 4/71 [copy-code-chunk]   
#> 
#> 
#> processing file: /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/tutorial.helpers/child_documents/copy_button.Rmd
#> 1/3 [change-code]      
#> 2/3                    
#> 3/3 [observe-transfers]
#> 5/71                     
#> 6/71 [unnamed-chunk-1]   
#> 7/71                     
#> 8/71 [unnamed-chunk-2]   
#> 9/71                     
#> 10/71 [info-section]      
#> 
#> 
#> processing file: /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/tutorial.helpers/child_documents/info_section.Rmd
#> 1/6        
#> 2/6 [name] 
#> 3/6        
#> 4/6 [email]
#> 5/6        
#> 6/6 [ID]   
#> 11/71                     
#> 12/71 [exercises-1]       
#> 13/71                     
#> 14/71 [exercises-1-hint-1]
#> 15/71                     
#> 16/71 [exercises-2]       
#> 17/71                     
#> 18/71 [exercises-2-hint-1]
#> 19/71                     
#> 20/71 [exercises-3]       
#> 21/71                     
#> 22/71 [exercises-3-hint-1]
#> 23/71                     
#> 24/71 [unnamed-chunk-3]   
#> 25/71                     
#> 26/71 [exercises-5]       
#> 27/71                     
#> 28/71 [exercises-5-test]  
#> 29/71                     
#> 30/71 [unnamed-chunk-4]   
#> 31/71                     
#> 32/71 [exercises-6]       
#> 33/71                     
#> 34/71 [exercises-7]       
#> 35/71                     
#> 36/71 [unnamed-chunk-5]   
#> 37/71                     
#> 38/71 [unnamed-chunk-6]   
#> 39/71                     
#> 40/71 [unnamed-chunk-7]   
#> 41/71                     
#> 42/71 [exercises-8]       
#> 43/71                     
#> 44/71 [exercises-9]       
#> 45/71                     
#> 46/71 [exercises-10]      
#> 47/71                     
#> 48/71 [exercises-11]      
#> 49/71                     
#> 50/71 [exercises-12]      
#> 51/71                     
#> 52/71 [exercises-13]      
#> 53/71                     
#> 54/71 [exercises-14]      
#> 55/71                     
#> 56/71 [exercises-15]      
#> 57/71                     
#> 58/71 [exercises-16]      
#> 59/71                     
#> 60/71 [exercises-17]      
#> 61/71                     
#> 62/71 [exercises-18]      
#> 63/71                     
#> 64/71 [exercises-19]      
#> 65/71                     
#> 66/71 [exercises-19-test] 
#> 67/71                     
#> 68/71 [exercises-20]      
#> 69/71                     
#> 70/71 [download-answers]  
#> 
#> 
#> processing file: /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/tutorial.helpers/child_documents/download_answers.Rmd
#> 1/6                   
#> 2/6 [minutes]         
#> 3/6                   
#> 4/6 [unnamed-chunk-47]
#> 5/6                   
#> 6/6 [unnamed-chunk-48]
#> 71/71                     
#> output file: /tmp/RtmpWuO7P0/tutorial.knit.md
#> /opt/hostedtoolcache/pandoc/3.1.11/x64/pandoc +RTS -K512m -RTS /tmp/RtmpWuO7P0/tutorial.knit.md --to html4 --from markdown+autolink_bare_uris+tex_math_single_backslash --output /tmp/RtmpWuO7P0/tutorial.html --lua-filter /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/rmarkdown/rmarkdown/lua/pagebreak.lua --lua-filter /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/rmarkdown/rmarkdown/lua/latex-div.lua --variable bs3=TRUE --standalone --section-divs --template /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/rmarkdown/rmd/h/default.html --no-highlight --variable highlightjs=1 --section-divs --reference-location=section --template /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/learnr/rmarkdown/templates/tutorial/resources/tutorial-format.htm --no-highlight --variable highlightjs=1 --variable progressive=true --variable allow-skip=true --variable learnr-version=0.11.5 --id-prefix section- --variable theme=cerulean --mathjax --variable 'mathjax-url=https://mathjax.rstudio.com/latest/MathJax.js?config=TeX-AMS-MML_HTMLorMML' --include-in-header /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/rmarkdown/rmd/h/shiny-header.html 
#> 
#> Output created: /tmp/RtmpWuO7P0/tutorial.html
#> Successfully rendered: /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/tutorial.helpers/tutorials/getting-started/tutorial.Rmd
#> Rendering: /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/tutorial.helpers/tutorials/introduction-to-r/tutorial.Rmd
#> 
#> 
#> processing file: tutorial.Rmd
#> 1/132                             
#> 2/132 [setup]                     
#> 3/132                             
#> 4/132 [copy-code-chunk]           
#> 
#> 
#> processing file: /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/tutorial.helpers/child_documents/copy_button.Rmd
#> 1/3 [change-code]      
#> 2/3                    
#> 3/3 [observe-transfers]
#> 5/132                             
#> 6/132 [info-section]              
#> 
#> 
#> processing file: /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/tutorial.helpers/child_documents/info_section.Rmd
#> 1/6        
#> 2/6 [name] 
#> 3/6        
#> 4/6 [email]
#> 5/6        
#> 6/6 [ID]   
#> 7/132                             
#> 8/132 [working-with-data-1]       
#> 9/132                             
#> 10/132 [working-with-data-1-hint-1]
#> 11/132                             
#> 12/132 [working-with-data-1-test]  
#> 13/132                             
#> 14/132 [working-with-data-2]       
#> 15/132                             
#> 16/132 [working-with-data-3]       
#> 17/132                             
#> 18/132 [working-with-data-3-test]  
#> 19/132                             
#> 20/132 [working-with-data-4]       
#> 21/132                             
#> 22/132 [working-with-data-4-test]  
#> 23/132                             
#> 24/132 [working-with-data-5]       
#> 25/132                             
#> 26/132 [working-with-data-5-test]  
#> 27/132                             
#> 28/132 [working-with-data-6]       
#> 29/132                             
#> 30/132 [working-with-data-6-test]  
#> 31/132                             
#> 32/132 [working-with-data-7]       
#> 33/132                             
#> 34/132 [working-with-data-7-test]  
#> 35/132                             
#> 36/132 [working-with-data-8]       
#> 37/132                             
#> 38/132 [working-with-data-8-test]  
#> 39/132                             
#> 40/132 [working-with-data-9]       
#> 41/132                             
#> 42/132 [working-with-data-10]      
#> 43/132                             
#> 44/132 [working-with-data-10-test] 
#> 45/132                             
#> 46/132 [working-with-data-11]      
#> 47/132                             
#> 48/132 [working-with-data-11-test] 
#> 49/132                             
#> 50/132 [working-with-data-12]      
#> 51/132                             
#> 52/132 [working-with-data-12-test] 
#> 53/132                             
#> 54/132 [working-with-data-13]      
#> 55/132                             
#> 56/132 [working-with-data-13-test] 
#> 57/132                             
#> 58/132 [pipes-and-plots-1]         
#> 59/132                             
#> 60/132 [pipes-and-plots-1-hint-1]  
#> 61/132                             
#> 62/132 [pipes-and-plots-1-test]    
#> 63/132                             
#> 64/132 [pipes-and-plots-2]         
#> 65/132                             
#> 66/132 [pipes-and-plots-2-hint-1]  
#> 67/132                             
#> 68/132 [pipes-and-plots-2-test]    
#> 69/132                             
#> 70/132 [pipes-and-plots-3]         
#> 71/132                             
#> 72/132 [pipes-and-plots-3-hint-1]  
#> 73/132                             
#> 74/132 [pipes-and-plots-3-test]    
#> 75/132                             
#> 76/132 [pipes-and-plots-4]         
#> 77/132                             
#> 78/132 [pipes-and-plots-4-hint-1]  
#> 79/132                             
#> 80/132 [pipes-and-plots-4-test]    
#> 81/132                             
#> 82/132 [pipes-and-plots-5]         
#> 83/132                             
#> 84/132 [pipes-and-plots-5-hint-1]  
#> 85/132                             
#> 86/132 [pipes-and-plots-5-test]    
#> 87/132                             
#> 88/132 [pipes-and-plots-6]         
#> 89/132                             
#> 90/132 [pipes-and-plots-6-hint-1]  
#> 91/132                             
#> 92/132 [pipes-and-plots-6-test]    
#> 93/132                             
#> 94/132 [pipes-and-plots-7]         
#> 95/132                             
#> 96/132 [pipes-and-plots-7-hint-1]  
#> 97/132                             
#> 98/132 [pipes-and-plots-7-test]    
#> 99/132                             
#> 100/132 [pipes-and-plots-8]         
#> 101/132                             
#> 102/132 [pipes-and-plots-8-hint-1]  
#> 103/132                             
#> 104/132 [pipes-and-plots-8-test]    
#> 105/132                             
#> 106/132 [pipes-and-plots-9]         
#> 107/132                             
#> 108/132 [pipes-and-plots-9-hint-1]  
#> 109/132                             
#> 110/132 [pipes-and-plots-9-test]    
#> 111/132                             
#> 112/132 [pipes-and-plots-10]        
#> 113/132                             
#> 114/132 [pipes-and-plots-10-hint-1] 
#> 115/132                             
#> 116/132 [pipes-and-plots-10-test]   
#> 117/132                             
#> 118/132 [pipes-and-plots-11]        
#> 119/132                             
#> 120/132 [pipes-and-plots-11-hint-1] 
#> 121/132                             
#> 122/132 [pipes-and-plots-11-test]   
#> 123/132                             
#> 124/132 [unnamed-chunk-1]           
#> 125/132                             
#> 126/132 [pipes-and-plots-12]        
#> 127/132                             
#> 128/132 [pipes-and-plots-12-hint-1] 
#> 129/132                             
#> 130/132 [pipes-and-plots-12-test]   
#> 131/132                             
#> 132/132 [download-answers]          
#> 
#> 
#> processing file: /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/tutorial.helpers/child_documents/download_answers.Rmd
#> 1/6                   
#> 2/6 [minutes]         
#> 3/6                   
#> 4/6 [unnamed-chunk-72]
#> 5/6                   
#> 6/6 [unnamed-chunk-73]
#> output file: /tmp/RtmpWuO7P0/tutorial.knit.md
#> /opt/hostedtoolcache/pandoc/3.1.11/x64/pandoc +RTS -K512m -RTS /tmp/RtmpWuO7P0/tutorial.knit.md --to html4 --from markdown+autolink_bare_uris+tex_math_single_backslash --output /tmp/RtmpWuO7P0/tutorial.html --lua-filter /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/rmarkdown/rmarkdown/lua/pagebreak.lua --lua-filter /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/rmarkdown/rmarkdown/lua/latex-div.lua --variable bs3=TRUE --standalone --section-divs --template /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/rmarkdown/rmd/h/default.html --no-highlight --variable highlightjs=1 --section-divs --reference-location=section --template /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/learnr/rmarkdown/templates/tutorial/resources/tutorial-format.htm --no-highlight --variable highlightjs=1 --variable progressive=true --variable allow-skip=true --variable learnr-version=0.11.5 --id-prefix section- --variable theme=cerulean --mathjax --variable 'mathjax-url=https://mathjax.rstudio.com/latest/MathJax.js?config=TeX-AMS-MML_HTMLorMML' --include-in-header /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/rmarkdown/rmd/h/shiny-header.html 
#> 
#> Output created: /tmp/RtmpWuO7P0/tutorial.html
#> Successfully rendered: /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/tutorial.helpers/tutorials/introduction-to-r/tutorial.Rmd
#> Successfully rendered 2 tutorial(s)