Confirm that a tutorial has the recommended components

There are three code components: the use of a copy-code button, an information request, and a download page. It is tricky to know where to store the "truth" of what these components should look like. For now, the truth is defined as the skeleton.Rmd which defines the template for creating a new tutorial.

All tutorials should also have library(learnr) and library(tutorial.helpers), both of which exist in the skeleton

Usage

check_tutorial_defaults(tutorial_paths)

Arguments

tutorial_paths

Character vector of the paths to the tutorials to be examined.

Value

No return value, called for side effects.

Examples

  check_tutorial_defaults(tutorial_paths = return_tutorial_paths("tutorial.helpers"))
#> NULL