Knit a set of tutorials

We define "testing" a tutorial as (successfully) running render() on it. This function renders all the tutorials provided in tutorial_paths. There is no check to see if the rendered file looks OK. If a tutorial fails to render, then an error will be generated which will propagate to the caller.

Usage

knit_tutorials(tutorial_paths)

Arguments

tutorial_paths

Character vector of the paths to the tutorials to be knitted.

Value

No return value, called for side effects.

Examples

  knit_tutorials(tutorial_paths = return_tutorial_paths("tutorial.helpers"))
#> Rendering: /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/tutorial.helpers/tutorials/getting-started/tutorial.Rmd
#> 
#> 
#> processing file: tutorial.Rmd
#> 1/75                     
#> 2/75 [setup]             
#> 3/75                     
#> 4/75 [copy-code-chunk]   
#> 
#> 
#> processing file: /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/tutorial.helpers/child_documents/copy_button.Rmd
#> 1/3 [change-code]      
#> 2/3                    
#> 3/3 [observe-transfers]
#> 5/75                     
#> 6/75 [unnamed-chunk-1]   
#> 7/75                     
#> 8/75 [unnamed-chunk-2]   
#> 9/75                     
#> 10/75 [info-section]      
#> 
#> 
#> processing file: /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/tutorial.helpers/child_documents/info_section.Rmd
#> 1/6        
#> 2/6 [name] 
#> 3/6        
#> 4/6 [email]
#> 5/6        
#> 6/6 [ID]   
#> 11/75                     
#> 12/75 [unnamed-chunk-3]   
#> 13/75                     
#> 14/75 [exercises-1]       
#> 15/75                     
#> 16/75 [exercises-1-hint-1]
#> 17/75                     
#> 18/75 [unnamed-chunk-4]   
#> 19/75                     
#> 20/75 [exercises-2]       
#> 21/75                     
#> 22/75 [exercises-2-hint-1]
#> 23/75                     
#> 24/75 [exercises-3]       
#> 25/75                     
#> 26/75 [exercises-3-hint-1]
#> 27/75                     
#> 28/75 [unnamed-chunk-5]   
#> 29/75                     
#> 30/75 [exercises-5]       
#> 31/75                     
#> 32/75 [unnamed-chunk-6]   
#> 33/75                     
#> 34/75 [exercises-6]       
#> 35/75                     
#> 36/75 [exercises-7]       
#> 37/75                     
#> 38/75 [unnamed-chunk-7]   
#> 39/75                     
#> 40/75 [unnamed-chunk-8]   
#> 41/75                     
#> 42/75 [exercises-8]       
#> 43/75                     
#> 44/75 [exercises-9]       
#> 45/75                     
#> 46/75 [exercises-10]      
#> 47/75                     
#> 48/75 [exercises-11]      
#> 49/75                     
#> 50/75 [exercises-12]      
#> 51/75                     
#> 52/75 [exercises-13]      
#> 53/75                     
#> 54/75 [exercises-14]      
#> 55/75                     
#> 56/75 [exercises-15]      
#> 57/75                     
#> 58/75 [exercises-16]      
#> 59/75                     
#> 60/75 [exercises-17]      
#> 61/75                     
#> 62/75 [exercises-18]      
#> 63/75                     
#> 64/75 [exercises-19]      
#> 65/75                     
#> 66/75 [exercises-20]      
#> 67/75                     
#> 68/75 [exercises-21]      
#> 69/75                     
#> 70/75 [exercises-21-test] 
#> 71/75                     
#> 72/75 [exercises-22]      
#> 73/75                     
#> 74/75 [download-answers]  
#> 
#> 
#> processing file: /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/tutorial.helpers/child_documents/download_answers.Rmd
#> 1/6                   
#> 2/6 [minutes]         
#> 3/6                   
#> 4/6 [unnamed-chunk-50]
#> 5/6                   
#> 6/6 [unnamed-chunk-51]
#> 75/75                     
#> output file: /tmp/RtmpcuRBQB/tutorial.knit.md
#> /opt/hostedtoolcache/pandoc/3.1.11/x64/pandoc +RTS -K512m -RTS /tmp/RtmpcuRBQB/tutorial.knit.md --to html4 --from markdown+autolink_bare_uris+tex_math_single_backslash --output /tmp/RtmpcuRBQB/tutorial.html --lua-filter /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/rmarkdown/rmarkdown/lua/pagebreak.lua --lua-filter /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/rmarkdown/rmarkdown/lua/latex-div.lua --variable bs3=TRUE --standalone --section-divs --template /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/rmarkdown/rmd/h/default.html --no-highlight --variable highlightjs=1 --section-divs --reference-location=section --template /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/learnr/rmarkdown/templates/tutorial/resources/tutorial-format.htm --no-highlight --variable highlightjs=1 --variable progressive=true --variable allow-skip=true --variable learnr-version=0.11.5 --id-prefix section- --variable theme=cerulean --mathjax --variable 'mathjax-url=https://mathjax.rstudio.com/latest/MathJax.js?config=TeX-AMS-MML_HTMLorMML' --include-in-header /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/rmarkdown/rmd/h/shiny-header.html 
#> 
#> Output created: /tmp/RtmpcuRBQB/tutorial.html
#> Successfully rendered: /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/tutorial.helpers/tutorials/getting-started/tutorial.Rmd
#> Rendering: /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/tutorial.helpers/tutorials/introduction-to-python/tutorial.Rmd
#> 
#> 
#> processing file: tutorial.Rmd
#> 1/8                   
#> 2/8 [setup]           
#> 3/8                   
#> 4/8 [copy-code-chunk] 
#> 
#> 
#> processing file: /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/tutorial.helpers/child_documents/copy_button.Rmd
#> 1/3 [change-code]      
#> 2/3                    
#> 3/3 [observe-transfers]
#> 5/8                   
#> 6/8 [info-section]    
#> 
#> 
#> processing file: /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/tutorial.helpers/child_documents/info_section.Rmd
#> 1/6        
#> 2/6 [name] 
#> 3/6        
#> 4/6 [email]
#> 5/6        
#> 6/6 [ID]   
#> 7/8                   
#> 8/8 [download-answers]
#> 
#> 
#> processing file: /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/tutorial.helpers/child_documents/download_answers.Rmd
#> 1/6                   
#> 2/6 [minutes]         
#> 3/6                   
#> 4/6 [unnamed-chunk-9] 
#> 5/6                   
#> 6/6 [unnamed-chunk-10]
#> output file: /tmp/RtmpcuRBQB/tutorial.knit.md
#> /opt/hostedtoolcache/pandoc/3.1.11/x64/pandoc +RTS -K512m -RTS /tmp/RtmpcuRBQB/tutorial.knit.md --to html4 --from markdown+autolink_bare_uris+tex_math_single_backslash --output /tmp/RtmpcuRBQB/tutorial.html --lua-filter /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/rmarkdown/rmarkdown/lua/pagebreak.lua --lua-filter /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/rmarkdown/rmarkdown/lua/latex-div.lua --variable bs3=TRUE --standalone --section-divs --template /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/rmarkdown/rmd/h/default.html --no-highlight --variable highlightjs=1 --section-divs --reference-location=section --template /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/learnr/rmarkdown/templates/tutorial/resources/tutorial-format.htm --no-highlight --variable highlightjs=1 --variable progressive=true --variable allow-skip=true --variable learnr-version=0.11.5 --id-prefix section- --variable theme=cerulean --mathjax --variable 'mathjax-url=https://mathjax.rstudio.com/latest/MathJax.js?config=TeX-AMS-MML_HTMLorMML' --include-in-header /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/rmarkdown/rmd/h/shiny-header.html 
#> 
#> Output created: /tmp/RtmpcuRBQB/tutorial.html
#> Successfully rendered: /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/tutorial.helpers/tutorials/introduction-to-python/tutorial.Rmd
#> Rendering: /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/tutorial.helpers/tutorials/introduction-to-r/tutorial.Rmd
#> 
#> 
#> processing file: tutorial.Rmd
#> 1/134                             
#> 2/134 [setup]                     
#> 3/134                             
#> 4/134 [copy-code-chunk]           
#> 
#> 
#> processing file: /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/tutorial.helpers/child_documents/copy_button.Rmd
#> 1/3 [change-code]      
#> 2/3                    
#> 3/3 [observe-transfers]
#> 5/134                             
#> 6/134 [info-section]              
#> 
#> 
#> processing file: /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/tutorial.helpers/child_documents/info_section.Rmd
#> 1/6        
#> 2/6 [name] 
#> 3/6        
#> 4/6 [email]
#> 5/6        
#> 6/6 [ID]   
#> 7/134                             
#> 8/134 [unnamed-chunk-1]           
#> 9/134                             
#> 10/134 [working-with-data-1]       
#> 11/134                             
#> 12/134 [working-with-data-1-hint-1]
#> 13/134                             
#> 14/134 [working-with-data-1-test]  
#> 15/134                             
#> 16/134 [working-with-data-2]       
#> 17/134                             
#> 18/134 [working-with-data-2-test]  
#> 19/134                             
#> 20/134 [working-with-data-3]       
#> 21/134                             
#> 22/134 [working-with-data-3-test]  
#> 23/134                             
#> 24/134 [working-with-data-4]       
#> 25/134                             
#> 26/134 [working-with-data-4-test]  
#> 27/134                             
#> 28/134 [working-with-data-5]       
#> 29/134                             
#> 30/134 [working-with-data-5-test]  
#> 31/134                             
#> 32/134 [working-with-data-6]       
#> 33/134                             
#> 34/134 [working-with-data-6-test]  
#> 35/134                             
#> 36/134 [working-with-data-7]       
#> 37/134                             
#> 38/134 [working-with-data-7-test]  
#> 39/134                             
#> 40/134 [working-with-data-8]       
#> 41/134                             
#> 42/134 [working-with-data-8-test]  
#> 43/134                             
#> 44/134 [working-with-data-9]       
#> 45/134                             
#> 46/134 [working-with-data-10]      
#> 47/134                             
#> 48/134 [working-with-data-10-test] 
#> 49/134                             
#> 50/134 [working-with-data-11]      
#> 51/134                             
#> 52/134 [working-with-data-11-test] 
#> 53/134                             
#> 54/134 [pipes-and-plots-1]         
#> 55/134                             
#> 56/134 [pipes-and-plots-1-hint-1]  
#> 57/134                             
#> 58/134 [pipes-and-plots-1-test]    
#> 59/134                             
#> 60/134 [pipes-and-plots-2]         
#> 61/134                             
#> 62/134 [pipes-and-plots-2-hint-1]  
#> 63/134                             
#> 64/134 [pipes-and-plots-2-test]    
#> 65/134                             
#> 66/134 [pipes-and-plots-3]         
#> 67/134                             
#> 68/134 [pipes-and-plots-3-hint-1]  
#> 69/134                             
#> 70/134 [pipes-and-plots-3-test]    
#> 71/134                             
#> 72/134 [pipes-and-plots-4]         
#> 73/134                             
#> 74/134 [pipes-and-plots-4-hint-1]  
#> 75/134                             
#> 76/134 [pipes-and-plots-4-test]    
#> 77/134                             
#> 78/134 [pipes-and-plots-5]         
#> 79/134                             
#> 80/134 [pipes-and-plots-5-hint-1]  
#> 81/134                             
#> 82/134 [pipes-and-plots-5-test]    
#> 83/134                             
#> 84/134 [pipes-and-plots-6]         
#> 85/134                             
#> 86/134 [pipes-and-plots-6-hint-1]  
#> 87/134                             
#> 88/134 [pipes-and-plots-6-test]    
#> 89/134                             
#> 90/134 [pipes-and-plots-7]         
#> 91/134                             
#> 92/134 [pipes-and-plots-7-hint-1]  
#> 93/134                             
#> 94/134 [pipes-and-plots-7-test]    
#> 95/134                             
#> 96/134 [pipes-and-plots-8]         
#> 97/134                             
#> 98/134 [pipes-and-plots-8-hint-1]  
#> 99/134                             
#> 100/134 [pipes-and-plots-8-test]    
#> 101/134                             
#> 102/134 [pipes-and-plots-9]         
#> 103/134                             
#> 104/134 [pipes-and-plots-9-hint-1]  
#> 105/134                             
#> 106/134 [pipes-and-plots-9-test]    
#> 107/134                             
#> 108/134 [pipes-and-plots-10]        
#> 109/134                             
#> 110/134 [pipes-and-plots-10-hint-1] 
#> 111/134                             
#> 112/134 [pipes-and-plots-10-test]   
#> 113/134                             
#> 114/134 [pipes-and-plots-11]        
#> 115/134                             
#> 116/134 [pipes-and-plots-11-hint-1] 
#> 117/134                             
#> 118/134 [pipes-and-plots-11-test]   
#> 119/134                             
#> 120/134 [pipes-and-plots-12]        
#> 121/134                             
#> 122/134 [pipes-and-plots-12-hint-1] 
#> 123/134                             
#> 124/134 [pipes-and-plots-12-test]   
#> 125/134                             
#> 126/134 [unnamed-chunk-2]           
#> 127/134                             
#> 128/134 [pipes-and-plots-13]        
#> 129/134                             
#> 130/134 [pipes-and-plots-13-hint-1] 
#> 131/134                             
#> 132/134 [pipes-and-plots-13-test]   
#> 133/134                             
#> 134/134 [download-answers]          
#> 
#> 
#> processing file: /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/tutorial.helpers/child_documents/download_answers.Rmd
#> 1/6                   
#> 2/6 [minutes]         
#> 3/6                   
#> 4/6 [unnamed-chunk-74]
#> 5/6                   
#> 6/6 [unnamed-chunk-75]
#> output file: /tmp/RtmpcuRBQB/tutorial.knit.md
#> /opt/hostedtoolcache/pandoc/3.1.11/x64/pandoc +RTS -K512m -RTS /tmp/RtmpcuRBQB/tutorial.knit.md --to html4 --from markdown+autolink_bare_uris+tex_math_single_backslash --output /tmp/RtmpcuRBQB/tutorial.html --lua-filter /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/rmarkdown/rmarkdown/lua/pagebreak.lua --lua-filter /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/rmarkdown/rmarkdown/lua/latex-div.lua --variable bs3=TRUE --standalone --section-divs --template /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/rmarkdown/rmd/h/default.html --no-highlight --variable highlightjs=1 --section-divs --reference-location=section --template /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/learnr/rmarkdown/templates/tutorial/resources/tutorial-format.htm --no-highlight --variable highlightjs=1 --variable progressive=true --variable allow-skip=true --variable learnr-version=0.11.5 --id-prefix section- --variable theme=cerulean --mathjax --variable 'mathjax-url=https://mathjax.rstudio.com/latest/MathJax.js?config=TeX-AMS-MML_HTMLorMML' --include-in-header /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/rmarkdown/rmd/h/shiny-header.html 
#> 
#> Output created: /tmp/RtmpcuRBQB/tutorial.html
#> Successfully rendered: /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/tutorial.helpers/tutorials/introduction-to-r/tutorial.Rmd
#> Successfully rendered 3 tutorial(s)