Knit a set of tutorials

We define "testing" a tutorial as (successfully) running render() on it. This function renders all the tutorials provided in tutorial_paths. There is no check to see if the rendered file looks OK. If a tutorial fails to render, then an error will be generated which will propagate to the caller.

Usage

knit_tutorials(tutorial_paths)

Arguments

tutorial_paths

Character vector of the paths to the tutorials to be knitted.

Value

No return value, called for side effects.

Examples

  knit_tutorials(tutorial_paths = return_tutorial_paths("tutorial.helpers"))
#> Rendering: /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/tutorial.helpers/tutorials/getting-started/tutorial.Rmd
#> 
#> 
#> processing file: tutorial.Rmd
#> 1/71                     
#> 2/71 [setup]             
#> 3/71                     
#> 4/71 [copy-code-chunk]   
#> 
#> 
#> processing file: /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/tutorial.helpers/child_documents/copy_button.Rmd
#> 1/3 [change-code]      
#> 2/3                    
#> 3/3 [observe-transfers]
#> 5/71                     
#> 6/71 [unnamed-chunk-1]   
#> 7/71                     
#> 8/71 [unnamed-chunk-2]   
#> 9/71                     
#> 10/71 [info-section]      
#> 
#> 
#> processing file: /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/tutorial.helpers/child_documents/info_section.Rmd
#> 1/6        
#> 2/6 [name] 
#> 3/6        
#> 4/6 [email]
#> 5/6        
#> 6/6 [ID]   
#> 11/71                     
#> 12/71 [unnamed-chunk-3]   
#> 13/71                     
#> 14/71 [exercises-1]       
#> 15/71                     
#> 16/71 [exercises-1-hint-1]
#> 17/71                     
#> 18/71 [unnamed-chunk-4]   
#> 19/71                     
#> 20/71 [exercises-2]       
#> 21/71                     
#> 22/71 [exercises-2-hint-1]
#> 23/71                     
#> 24/71 [exercises-3]       
#> 25/71                     
#> 26/71 [exercises-3-hint-1]
#> 27/71                     
#> 28/71 [unnamed-chunk-5]   
#> 29/71                     
#> 30/71 [exercises-5]       
#> 31/71                     
#> 32/71 [unnamed-chunk-6]   
#> 33/71                     
#> 34/71 [exercises-6]       
#> 35/71                     
#> 36/71 [exercises-7]       
#> 37/71                     
#> 38/71 [unnamed-chunk-7]   
#> 39/71                     
#> 40/71 [unnamed-chunk-8]   
#> 41/71                     
#> 42/71 [exercises-8]       
#> 43/71                     
#> 44/71 [exercises-9]       
#> 45/71                     
#> 46/71 [exercises-10]      
#> 47/71                     
#> 48/71 [exercises-11]      
#> 49/71                     
#> 50/71 [exercises-12]      
#> 51/71                     
#> 52/71 [exercises-13]      
#> 53/71                     
#> 54/71 [exercises-14]      
#> 55/71                     
#> 56/71 [exercises-15]      
#> 57/71                     
#> 58/71 [exercises-16]      
#> 59/71                     
#> 60/71 [exercises-17]      
#> 61/71                     
#> 62/71 [exercises-18]      
#> 63/71                     
#> 64/71 [exercises-19]      
#> 65/71                     
#> 66/71 [exercises-19-test] 
#> 67/71                     
#> 68/71 [exercises-20]      
#> 69/71                     
#> 70/71 [download-answers]  
#> 
#> 
#> processing file: /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/tutorial.helpers/child_documents/download_answers.Rmd
#> 1/6                   
#> 2/6 [minutes]         
#> 3/6                   
#> 4/6 [unnamed-chunk-48]
#> 5/6                   
#> 6/6 [unnamed-chunk-49]
#> 71/71                     
#> output file: /tmp/Rtmp1S0zfS/tutorial.knit.md
#> /opt/hostedtoolcache/pandoc/3.1.11/x64/pandoc +RTS -K512m -RTS /tmp/Rtmp1S0zfS/tutorial.knit.md --to html4 --from markdown+autolink_bare_uris+tex_math_single_backslash --output /tmp/Rtmp1S0zfS/tutorial.html --lua-filter /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/rmarkdown/rmarkdown/lua/pagebreak.lua --lua-filter /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/rmarkdown/rmarkdown/lua/latex-div.lua --variable bs3=TRUE --standalone --section-divs --template /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/rmarkdown/rmd/h/default.html --no-highlight --variable highlightjs=1 --section-divs --reference-location=section --template /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/learnr/rmarkdown/templates/tutorial/resources/tutorial-format.htm --no-highlight --variable highlightjs=1 --variable progressive=true --variable allow-skip=true --variable learnr-version=0.11.5 --id-prefix section- --variable theme=cerulean --mathjax --variable 'mathjax-url=https://mathjax.rstudio.com/latest/MathJax.js?config=TeX-AMS-MML_HTMLorMML' --include-in-header /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/rmarkdown/rmd/h/shiny-header.html 
#> 
#> Output created: /tmp/Rtmp1S0zfS/tutorial.html
#> Successfully rendered: /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/tutorial.helpers/tutorials/getting-started/tutorial.Rmd
#> Rendering: /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/tutorial.helpers/tutorials/introduction-to-python/tutorial.Rmd
#> 
#> 
#> processing file: tutorial.Rmd
#> 1/8                   
#> 2/8 [setup]           
#> 3/8                   
#> 4/8 [copy-code-chunk] 
#> 
#> 
#> processing file: /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/tutorial.helpers/child_documents/copy_button.Rmd
#> 1/3 [change-code]      
#> 2/3                    
#> 3/3 [observe-transfers]
#> 5/8                   
#> 6/8 [info-section]    
#> 
#> 
#> processing file: /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/tutorial.helpers/child_documents/info_section.Rmd
#> 1/6        
#> 2/6 [name] 
#> 3/6        
#> 4/6 [email]
#> 5/6        
#> 6/6 [ID]   
#> 7/8                   
#> 8/8 [download-answers]
#> 
#> 
#> processing file: /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/tutorial.helpers/child_documents/download_answers.Rmd
#> 1/6                   
#> 2/6 [minutes]         
#> 3/6                   
#> 4/6 [unnamed-chunk-9] 
#> 5/6                   
#> 6/6 [unnamed-chunk-10]
#> output file: /tmp/Rtmp1S0zfS/tutorial.knit.md
#> /opt/hostedtoolcache/pandoc/3.1.11/x64/pandoc +RTS -K512m -RTS /tmp/Rtmp1S0zfS/tutorial.knit.md --to html4 --from markdown+autolink_bare_uris+tex_math_single_backslash --output /tmp/Rtmp1S0zfS/tutorial.html --lua-filter /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/rmarkdown/rmarkdown/lua/pagebreak.lua --lua-filter /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/rmarkdown/rmarkdown/lua/latex-div.lua --variable bs3=TRUE --standalone --section-divs --template /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/rmarkdown/rmd/h/default.html --no-highlight --variable highlightjs=1 --section-divs --reference-location=section --template /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/learnr/rmarkdown/templates/tutorial/resources/tutorial-format.htm --no-highlight --variable highlightjs=1 --variable progressive=true --variable allow-skip=true --variable learnr-version=0.11.5 --id-prefix section- --variable theme=cerulean --mathjax --variable 'mathjax-url=https://mathjax.rstudio.com/latest/MathJax.js?config=TeX-AMS-MML_HTMLorMML' --include-in-header /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/rmarkdown/rmd/h/shiny-header.html 
#> 
#> Output created: /tmp/Rtmp1S0zfS/tutorial.html
#> Successfully rendered: /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/tutorial.helpers/tutorials/introduction-to-python/tutorial.Rmd
#> Rendering: /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/tutorial.helpers/tutorials/introduction-to-r/tutorial.Rmd
#> 
#> 
#> processing file: tutorial.Rmd
#> 1/138                             
#> 2/138 [setup]                     
#> 3/138                             
#> 4/138 [copy-code-chunk]           
#> 
#> 
#> processing file: /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/tutorial.helpers/child_documents/copy_button.Rmd
#> 1/3 [change-code]      
#> 2/3                    
#> 3/3 [observe-transfers]
#> 5/138                             
#> 6/138 [info-section]              
#> 
#> 
#> processing file: /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/tutorial.helpers/child_documents/info_section.Rmd
#> 1/6        
#> 2/6 [name] 
#> 3/6        
#> 4/6 [email]
#> 5/6        
#> 6/6 [ID]   
#> 7/138                             
#> 8/138 [working-with-data-1]       
#> 9/138                             
#> 10/138 [working-with-data-1-hint-1]
#> 11/138                             
#> 12/138 [working-with-data-1-test]  
#> 13/138                             
#> 14/138 [working-with-data-2]       
#> 15/138                             
#> 16/138 [working-with-data-2-test]  
#> 17/138                             
#> 18/138 [working-with-data-3]       
#> 19/138                             
#> 20/138 [working-with-data-3-test]  
#> 21/138                             
#> 22/138 [working-with-data-4]       
#> 23/138                             
#> 24/138 [working-with-data-4-test]  
#> 25/138                             
#> 26/138 [working-with-data-5]       
#> 27/138                             
#> 28/138 [working-with-data-5-test]  
#> 29/138                             
#> 30/138 [working-with-data-6]       
#> 31/138                             
#> 32/138 [working-with-data-6-test]  
#> 33/138                             
#> 34/138 [working-with-data-7]       
#> 35/138                             
#> 36/138 [working-with-data-7-test]  
#> 37/138                             
#> 38/138 [working-with-data-8]       
#> 39/138                             
#> 40/138 [working-with-data-8-test]  
#> 41/138                             
#> 42/138 [working-with-data-9]       
#> 43/138                             
#> 44/138 [working-with-data-10]      
#> 45/138                             
#> 46/138 [working-with-data-10-test] 
#> 47/138                             
#> 48/138 [working-with-data-11]      
#> 49/138                             
#> 50/138 [working-with-data-11-test] 
#> 51/138                             
#> 52/138 [working-with-data-12]      
#> 53/138                             
#> 54/138 [working-with-data-12-test] 
#> 55/138                             
#> 56/138 [working-with-data-13]      
#> 57/138                             
#> 58/138 [working-with-data-13-test] 
#> 59/138                             
#> 60/138 [working-with-data-14]      
#> 61/138                             
#> 62/138 [working-with-data-14-test] 
#> 63/138                             
#> 64/138 [pipes-and-plots-1]         
#> 65/138                             
#> 66/138 [pipes-and-plots-1-hint-1]  
#> 67/138                             
#> 68/138 [pipes-and-plots-1-test]    
#> 69/138                             
#> 70/138 [pipes-and-plots-2]         
#> 71/138                             
#> 72/138 [pipes-and-plots-2-hint-1]  
#> 73/138                             
#> 74/138 [pipes-and-plots-2-test]    
#> 75/138                             
#> 76/138 [pipes-and-plots-3]         
#> 77/138                             
#> 78/138 [pipes-and-plots-3-hint-1]  
#> 79/138                             
#> 80/138 [pipes-and-plots-3-test]    
#> 81/138                             
#> 82/138 [pipes-and-plots-4]         
#> 83/138                             
#> 84/138 [pipes-and-plots-4-hint-1]  
#> 85/138                             
#> 86/138 [pipes-and-plots-4-test]    
#> 87/138                             
#> 88/138 [pipes-and-plots-5]         
#> 89/138                             
#> 90/138 [pipes-and-plots-5-hint-1]  
#> 91/138                             
#> 92/138 [pipes-and-plots-5-test]    
#> 93/138                             
#> 94/138 [pipes-and-plots-6]         
#> 95/138                             
#> 96/138 [pipes-and-plots-6-hint-1]  
#> 97/138                             
#> 98/138 [pipes-and-plots-6-test]    
#> 99/138                             
#> 100/138 [pipes-and-plots-7]         
#> 101/138                             
#> 102/138 [pipes-and-plots-7-hint-1]  
#> 103/138                             
#> 104/138 [pipes-and-plots-7-test]    
#> 105/138                             
#> 106/138 [pipes-and-plots-8]         
#> 107/138                             
#> 108/138 [pipes-and-plots-8-hint-1]  
#> 109/138                             
#> 110/138 [pipes-and-plots-8-test]    
#> 111/138                             
#> 112/138 [pipes-and-plots-9]         
#> 113/138                             
#> 114/138 [pipes-and-plots-9-hint-1]  
#> 115/138                             
#> 116/138 [pipes-and-plots-9-test]    
#> 117/138                             
#> 118/138 [pipes-and-plots-10]        
#> 119/138                             
#> 120/138 [pipes-and-plots-10-hint-1] 
#> 121/138                             
#> 122/138 [pipes-and-plots-10-test]   
#> 123/138                             
#> 124/138 [pipes-and-plots-11]        
#> 125/138                             
#> 126/138 [pipes-and-plots-11-hint-1] 
#> 127/138                             
#> 128/138 [pipes-and-plots-11-test]   
#> 129/138                             
#> 130/138 [unnamed-chunk-1]           
#> 131/138                             
#> 132/138 [pipes-and-plots-12]        
#> 133/138                             
#> 134/138 [pipes-and-plots-12-hint-1] 
#> 135/138                             
#> 136/138 [pipes-and-plots-12-test]   
#> 137/138                             
#> 138/138 [download-answers]          
#> 
#> 
#> processing file: /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/tutorial.helpers/child_documents/download_answers.Rmd
#> 1/6                   
#> 2/6 [minutes]         
#> 3/6                   
#> 4/6 [unnamed-chunk-75]
#> 5/6                   
#> 6/6 [unnamed-chunk-76]
#> output file: /tmp/Rtmp1S0zfS/tutorial.knit.md
#> /opt/hostedtoolcache/pandoc/3.1.11/x64/pandoc +RTS -K512m -RTS /tmp/Rtmp1S0zfS/tutorial.knit.md --to html4 --from markdown+autolink_bare_uris+tex_math_single_backslash --output /tmp/Rtmp1S0zfS/tutorial.html --lua-filter /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/rmarkdown/rmarkdown/lua/pagebreak.lua --lua-filter /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/rmarkdown/rmarkdown/lua/latex-div.lua --variable bs3=TRUE --standalone --section-divs --template /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/rmarkdown/rmd/h/default.html --no-highlight --variable highlightjs=1 --section-divs --reference-location=section --template /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/learnr/rmarkdown/templates/tutorial/resources/tutorial-format.htm --no-highlight --variable highlightjs=1 --variable progressive=true --variable allow-skip=true --variable learnr-version=0.11.5 --id-prefix section- --variable theme=cerulean --mathjax --variable 'mathjax-url=https://mathjax.rstudio.com/latest/MathJax.js?config=TeX-AMS-MML_HTMLorMML' --include-in-header /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/rmarkdown/rmd/h/shiny-header.html 
#> 
#> Output created: /tmp/Rtmp1S0zfS/tutorial.html
#> Successfully rendered: /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/tutorial.helpers/tutorials/introduction-to-r/tutorial.Rmd
#> Successfully rendered 3 tutorial(s)