Knit a set of tutorials

We define "testing" a tutorial as (successfully) running render() on it. This function renders all the tutorials provided in tutorial_paths. There is no check to see if the rendered file looks OK. If a tutorial fails to render, then (we assume!) an error will be generated which will then filter up to our testing rig.

Usage

knit_tutorials(tutorial_paths)

Arguments

tutorial_paths

Character vector of the paths to the tutorials to be knitted.

Value

No return value, called for side effects.

Examples

  knit_tutorials(tutorial_paths = return_tutorial_paths("tutorial.helpers"))
#> 
#> 
#> processing file: tutorial.Rmd
#>   |                                                |                                        |   0%  |                                                |.                                       |   3%                                |                                                |..                                      |   6% (setup)                        |                                                |....                                    |   9%                                |                                                |.....                                   |  12% (copy-code-chunk)            
#> 
#> 
#> processing file: /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/R-4.3/x86_64-pc-linux-gnu/tutorial.helpers/child_documents/copy_button.Rmd
#>   |                                                          |                                                  |   0%  |                                                          |.................                                 |  33% (change-code)        |                                                          |.................................                 |  67%                      |                                                          |..................................................| 100% (observe-transfers)                                                                                                              
#>   |                                                |......                                  |  15%                                |                                                |.......                                 |  18% (info-section)               
#> 
#> 
#> processing file: /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/R-4.3/x86_64-pc-linux-gnu/tutorial.helpers/child_documents/info_section.Rmd
#>   |                                                          |                                                  |   0%  |                                                          |............                                      |  25%                      |                                                          |.........................                         |  50% (information-name)   |                                                          |......................................            |  75%                      |                                                          |..................................................| 100% (information-email)                                                                                                              
#>   |                                                |........                                |  21%                                |                                                |..........                              |  24% (tutorial-exercises-1)         |                                                |...........                             |  27%                                |                                                |............                            |  30% (tutorial-exercises-1-hint-1)  |                                                |.............                           |  33%                                |                                                |...............                         |  36% (tutorial-exercises-2)         |                                                |................                        |  39%                                |                                                |.................                       |  42% (tutorial-exercises-2-hint-1)  |                                                |..................                      |  45%                                |                                                |...................                     |  48% (tutorial-exercises-3)         |                                                |.....................                   |  52%                                |                                                |......................                  |  55% (tutorial-exercises-4)         |                                                |.......................                 |  58%                                |                                                |........................                |  61% (tutorial-exercises-4-hint-1)  |                                                |.........................               |  64%                                |                                                |...........................             |  67% (unnamed-chunk-1)              |                                                |............................            |  70%                                |                                                |.............................           |  73% (tutorial-exercises-6)         |                                                |..............................          |  76%                                |                                                |................................        |  79% (tutorial-exercises-7)         |                                                |.................................       |  82%                                |                                                |..................................      |  85% (tutorial-exercises-8)         |                                                |...................................     |  88%                                |                                                |....................................    |  91% (tutorial-exercises-9)         |                                                |......................................  |  94%                                |                                                |....................................... |  97% (download-answers)           
#> 
#> 
#> processing file: /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/R-4.3/x86_64-pc-linux-gnu/tutorial.helpers/child_documents/download_answers.Rmd
#>   |                                                         |                                                 |   0%  |                                                         |........                                         |  17%                       |                                                         |................                                 |  33% (download-answers-1)  |                                                         |........................                         |  50%                       |                                                         |.................................                |  67% (unnamed-chunk-2)     |                                                         |.........................................        |  83%                       |                                                         |.................................................| 100% (unnamed-chunk-3)                                                                                                                  
#>   |                                                |........................................| 100%                                                                                                                                                      
#> output file: /tmp/RtmparewLL/tutorial.knit.md
#> /usr/bin/pandoc +RTS -K512m -RTS /tmp/RtmparewLL/tutorial.knit.md --to html4 --from markdown+autolink_bare_uris+tex_math_single_backslash --output /tmp/RtmparewLL/tutorial.html --lua-filter /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/R-4.3/x86_64-pc-linux-gnu/rmarkdown/rmarkdown/lua/pagebreak.lua --lua-filter /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/R-4.3/x86_64-pc-linux-gnu/rmarkdown/rmarkdown/lua/latex-div.lua --variable bs3=TRUE --standalone --section-divs --template /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/R-4.3/x86_64-pc-linux-gnu/rmarkdown/rmd/h/default.html --no-highlight --variable highlightjs=1 --section-divs --reference-location=section --template /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/R-4.3/x86_64-pc-linux-gnu/learnr/rmarkdown/templates/tutorial/resources/tutorial-format.htm --no-highlight --variable highlightjs=1 --variable progressive=true --variable allow-skip=true --variable learnr-version=0.11.3 --id-prefix section- --variable theme=cerulean --mathjax --variable 'mathjax-url=https://mathjax.rstudio.com/latest/MathJax.js?config=TeX-AMS-MML_HTMLorMML' --include-in-header /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/R-4.3/x86_64-pc-linux-gnu/rmarkdown/rmd/h/shiny-header.html 
#> 
#> Output created: /tmp/RtmparewLL/tutorial.html
#> ── Skip (???): Rendering /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/R-4.3/x86_64-pc-linux-gnu/tutorial.helpers/tutorials/getting-started/tutorial.Rmd ──
#> Reason: empty test
#> 
#> NULL