Knit a set of tutorials

We define "testing" a tutorial as (successfully) running render() on it. This function renders all the tutorials provided in tutorial_paths. There is no check to see if the rendered file looks OK. If a tutorial fails to render, then (we assume!) an error will be generated which will then filter up to our testing rig.

Usage

knit_tutorials(tutorial_paths)

Arguments

tutorial_paths

Character vector of the paths to the tutorials to be knitted.

Value

No return value, called for side effects.

Examples

  knit_tutorials(tutorial_paths = return_tutorial_paths("tutorial.helpers"))
#> 
#> 
#> processing file: tutorial.Rmd
#> 1/31                     
#> 2/31 [setup]             
#> 3/31                     
#> 4/31 [copy-code-chunk]   
#> 
#> 
#> processing file: /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-jammy/R-4.4/x86_64-pc-linux-gnu/tutorial.helpers/child_documents/copy_button.Rmd
#> 1/3 [change-code]      
#> 2/3                    
#> 3/3 [observe-transfers]
#> 5/31                     
#> 6/31 [info-section]      
#> 
#> 
#> processing file: /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-jammy/R-4.4/x86_64-pc-linux-gnu/tutorial.helpers/child_documents/info_section.Rmd
#> 1/6                    
#> 2/6 [information-name] 
#> 3/6                    
#> 4/6 [information-email]
#> 5/6                    
#> 6/6 [information-id]   
#> 7/31                     
#> 8/31 [exercises-1]       
#> 9/31                     
#> 10/31 [exercises-1-hint-1]
#> 11/31                     
#> 12/31 [exercises-2]       
#> 13/31                     
#> 14/31 [exercises-2-hint-1]
#> 15/31                     
#> 16/31 [exercises-3]       
#> 17/31                     
#> 18/31 [exercises-3-hint-1]
#> 19/31                     
#> 20/31 [unnamed-chunk-1]   
#> 21/31                     
#> 22/31 [exercises-5]       
#> 23/31                     
#> 24/31 [exercises-6]       
#> 25/31                     
#> 26/31 [exercises-7]       
#> 27/31                     
#> 28/31 [exercises-8]       
#> 29/31                     
#> 30/31 [download-answers]  
#> 
#> 
#> processing file: /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-jammy/R-4.4/x86_64-pc-linux-gnu/tutorial.helpers/child_documents/download_answers.Rmd
#> 1/6                   
#> 2/6 [download-answers]
#> 3/6                   
#> 4/6 [unnamed-chunk-21]
#> 5/6                   
#> 6/6 [unnamed-chunk-22]
#> 31/31                     
#> output file: /tmp/RtmpotitA6/tutorial.knit.md
#> /opt/hostedtoolcache/pandoc/3.1.11/x64/pandoc +RTS -K512m -RTS /tmp/RtmpotitA6/tutorial.knit.md --to html4 --from markdown+autolink_bare_uris+tex_math_single_backslash --output /tmp/RtmpotitA6/tutorial.html --lua-filter /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-jammy/R-4.4/x86_64-pc-linux-gnu/rmarkdown/rmarkdown/lua/pagebreak.lua --lua-filter /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-jammy/R-4.4/x86_64-pc-linux-gnu/rmarkdown/rmarkdown/lua/latex-div.lua --variable bs3=TRUE --standalone --section-divs --template /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-jammy/R-4.4/x86_64-pc-linux-gnu/rmarkdown/rmd/h/default.html --no-highlight --variable highlightjs=1 --section-divs --reference-location=section --template /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-jammy/R-4.4/x86_64-pc-linux-gnu/learnr/rmarkdown/templates/tutorial/resources/tutorial-format.htm --no-highlight --variable highlightjs=1 --variable progressive=true --variable allow-skip=true --variable learnr-version=0.11.5 --id-prefix section- --variable theme=cerulean --mathjax --variable 'mathjax-url=https://mathjax.rstudio.com/latest/MathJax.js?config=TeX-AMS-MML_HTMLorMML' --include-in-header /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-jammy/R-4.4/x86_64-pc-linux-gnu/rmarkdown/rmd/h/shiny-header.html 
#> 
#> Output created: /tmp/RtmpotitA6/tutorial.html
#> -- Skip: Rendering /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-jammy/R-4.4/x86_64-pc-linux-gnu/tutorial.helpers/tutorials/getting-started/tutorial.Rmd --
#> Reason: empty test
#> 
#> NULL