Knit a set of tutorials

We define "testing" a tutorial as (successfully) running render() on it. This function renders all the tutorials provided in tutorial_paths. There is no check to see if the rendered file looks OK. If a tutorial fails to render, then (we assume!) an error will be generated which will then filter up to our testing rig.

Usage

knit_tutorials(tutorial_paths)

Arguments

tutorial_paths

Character vector of the paths to the tutorials to be knitted.

Value

No return value, called for side effects.

Examples

  knit_tutorials(tutorial_paths = return_tutorial_paths("tutorial.helpers"))
#> 
#> 
#> processing file: tutorial.Rmd
#> 1/33                              
#> 2/33 [setup]                      
#> 3/33                              
#> 4/33 [copy-code-chunk]            
#> 
#> 
#> processing file: /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/R-4.3/x86_64-pc-linux-gnu/tutorial.helpers/child_documents/copy_button.Rmd
#> 1/3 [change-code]      
#> 2/3                    
#> 3/3 [observe-transfers]
#> 5/33                              
#> 6/33 [info-section]               
#> 
#> 
#> processing file: /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/R-4.3/x86_64-pc-linux-gnu/tutorial.helpers/child_documents/info_section.Rmd
#> 1/4                    
#> 2/4 [information-name] 
#> 3/4                    
#> 4/4 [information-email]
#> 7/33                              
#> 8/33 [tutorial-exercises-1]       
#> 9/33                              
#> 10/33 [tutorial-exercises-1-hint-1]
#> 11/33                              
#> 12/33 [tutorial-exercises-2]       
#> 13/33                              
#> 14/33 [tutorial-exercises-2-hint-1]
#> 15/33                              
#> 16/33 [tutorial-exercises-3]       
#> 17/33                              
#> 18/33 [tutorial-exercises-4]       
#> 19/33                              
#> 20/33 [tutorial-exercises-4-hint-1]
#> 21/33                              
#> 22/33 [unnamed-chunk-1]            
#> 23/33                              
#> 24/33 [tutorial-exercises-6]       
#> 25/33                              
#> 26/33 [tutorial-exercises-7]       
#> 27/33                              
#> 28/33 [tutorial-exercises-8]       
#> 29/33                              
#> 30/33 [tutorial-exercises-9]       
#> 31/33                              
#> 32/33 [download-answers]           
#> 
#> 
#> processing file: /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/R-4.3/x86_64-pc-linux-gnu/tutorial.helpers/child_documents/download_answers.Rmd
#> 1/6                     
#> 2/6 [download-answers-1]
#> 3/6                     
#> 4/6 [unnamed-chunk-21]  
#> 5/6                     
#> 6/6 [unnamed-chunk-22]  
#> 33/33                              
#> output file: /tmp/RtmpvA4DVE/tutorial.knit.md
#> /usr/bin/pandoc +RTS -K512m -RTS /tmp/RtmpvA4DVE/tutorial.knit.md --to html4 --from markdown+autolink_bare_uris+tex_math_single_backslash --output /tmp/RtmpvA4DVE/tutorial.html --lua-filter /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/R-4.3/x86_64-pc-linux-gnu/rmarkdown/rmarkdown/lua/pagebreak.lua --lua-filter /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/R-4.3/x86_64-pc-linux-gnu/rmarkdown/rmarkdown/lua/latex-div.lua --variable bs3=TRUE --standalone --section-divs --template /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/R-4.3/x86_64-pc-linux-gnu/rmarkdown/rmd/h/default.html --no-highlight --variable highlightjs=1 --section-divs --reference-location=section --template /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/R-4.3/x86_64-pc-linux-gnu/learnr/rmarkdown/templates/tutorial/resources/tutorial-format.htm --no-highlight --variable highlightjs=1 --variable progressive=true --variable allow-skip=true --variable learnr-version=0.11.5 --id-prefix section- --variable theme=cerulean --mathjax --variable 'mathjax-url=https://mathjax.rstudio.com/latest/MathJax.js?config=TeX-AMS-MML_HTMLorMML' --include-in-header /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/R-4.3/x86_64-pc-linux-gnu/rmarkdown/rmd/h/shiny-header.html 
#> 
#> Output created: /tmp/RtmpvA4DVE/tutorial.html
#> -- Skip: Rendering /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/R-4.3/x86_64-pc-linux-gnu/tutorial.helpers/tutorials/getting-started/tutorial.Rmd --
#> Reason: empty test
#> 
#> 
#> 
#> processing file: tutorial.Rmd
#> 1/35                              
#> 2/35 [setup]                      
#> 3/35                              
#> 4/35 [copy-code-chunk]            
#> 
#> 
#> processing file: /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/R-4.3/x86_64-pc-linux-gnu/tutorial.helpers/child_documents/copy_button.Rmd
#> 1/3 [change-code]      
#> 2/3                    
#> 3/3 [observe-transfers]
#> 5/35                              
#> 6/35 [info-section]               
#> 
#> 
#> processing file: /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/R-4.3/x86_64-pc-linux-gnu/tutorial.helpers/child_documents/info_section.Rmd
#> 1/4                    
#> 2/4 [information-name] 
#> 3/4                    
#> 4/4 [information-email]
#> 7/35                              
#> 8/35 [tutorial-exercises-1]       
#> 9/35                              
#> 10/35 [tutorial-exercises-1-hint-1]
#> 11/35                              
#> 12/35 [tutorial-exercises-2]       
#> 13/35                              
#> 14/35 [tutorial-exercises-2-hint-1]
#> 15/35                              
#> 16/35 [tutorial-exercises-3a]      
#> 17/35                              
#> 18/35 [tutorial-exercises-3b]      
#> 19/35                              
#> 20/35 [tutorial-exercises-4]       
#> 21/35                              
#> 22/35 [tutorial-exercises-4-hint-1]
#> 23/35                              
#> 24/35 [unnamed-chunk-1]            
#> 25/35                              
#> 26/35 [tutorial-exercises-6]       
#> 27/35                              
#> 28/35 [tutorial-exercises-7]       
#> 29/35                              
#> 30/35 [tutorial-exercises-8]       
#> 31/35                              
#> 32/35 [tutorial-exercises-9]       
#> 33/35                              
#> 34/35 [download-answers]           
#> 
#> 
#> processing file: /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/R-4.3/x86_64-pc-linux-gnu/tutorial.helpers/child_documents/download_answers.Rmd
#> 1/6                     
#> 2/6 [download-answers-1]
#> 3/6                     
#> 4/6 [unnamed-chunk-22]  
#> 5/6                     
#> 6/6 [unnamed-chunk-23]  
#> 35/35                              
#> output file: /tmp/RtmpvA4DVE/tutorial.knit.md
#> /usr/bin/pandoc +RTS -K512m -RTS /tmp/RtmpvA4DVE/tutorial.knit.md --to html4 --from markdown+autolink_bare_uris+tex_math_single_backslash --output /tmp/RtmpvA4DVE/tutorial.html --lua-filter /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/R-4.3/x86_64-pc-linux-gnu/rmarkdown/rmarkdown/lua/pagebreak.lua --lua-filter /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/R-4.3/x86_64-pc-linux-gnu/rmarkdown/rmarkdown/lua/latex-div.lua --variable bs3=TRUE --standalone --section-divs --template /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/R-4.3/x86_64-pc-linux-gnu/rmarkdown/rmd/h/default.html --no-highlight --variable highlightjs=1 --section-divs --reference-location=section --template /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/R-4.3/x86_64-pc-linux-gnu/learnr/rmarkdown/templates/tutorial/resources/tutorial-format.htm --no-highlight --variable highlightjs=1 --variable progressive=true --variable allow-skip=true --variable learnr-version=0.11.5 --id-prefix section- --variable theme=cerulean --mathjax --variable 'mathjax-url=https://mathjax.rstudio.com/latest/MathJax.js?config=TeX-AMS-MML_HTMLorMML' --include-in-header /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/R-4.3/x86_64-pc-linux-gnu/rmarkdown/rmd/h/shiny-header.html 
#> 
#> Output created: /tmp/RtmpvA4DVE/tutorial.html
#> -- Skip: Rendering /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/R-4.3/x86_64-pc-linux-gnu/tutorial.helpers/tutorials/getting-started_es/tutorial.Rmd --
#> Reason: empty test
#> 
#> NULL