Knit a set of tutorials

We define "testing" a tutorial as (successfully) running render() on it. This function renders all the tutorials provided in tutorial_paths. There is no check to see if the rendered file looks OK. If a tutorial fails to render, then (we assume!) an error will be generated which will then filter up to our testing rig.

Usage

knit_tutorials(tutorial_paths)

Arguments

tutorial_paths

Character vector of the paths to the tutorials to be knitted.

Value

No return value, called for side effects.

Examples

  knit_tutorials(tutorial_paths = return_tutorial_paths("tutorial.helpers"))
#> 
#> 
#> processing file: tutorial.Rmd
#> 1/53                     
#> 2/53 [setup]             
#> 3/53                     
#> 4/53 [copy-code-chunk]   
#> 
#> 
#> processing file: /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/tutorial.helpers/child_documents/copy_button.Rmd
#> 1/3 [change-code]      
#> 2/3                    
#> 3/3 [observe-transfers]
#> 5/53                     
#> 6/53 [unnamed-chunk-1]   
#> 7/53                     
#> 8/53 [unnamed-chunk-2]   
#> 9/53                     
#> 10/53 [info-section]      
#> 
#> 
#> processing file: /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/tutorial.helpers/child_documents/info_section.Rmd
#> 1/6        
#> 2/6 [name] 
#> 3/6        
#> 4/6 [email]
#> 5/6        
#> 6/6 [ID]   
#> 11/53                     
#> 12/53 [unnamed-chunk-3]   
#> 13/53                     
#> 14/53 [exercises-1]       
#> 15/53                     
#> 16/53 [exercises-1-hint-1]
#> 17/53                     
#> 18/53 [unnamed-chunk-4]   
#> 19/53                     
#> 20/53 [exercises-2]       
#> 21/53                     
#> 22/53 [exercises-2-hint-1]
#> 23/53                     
#> 24/53 [exercises-3]       
#> 25/53                     
#> 26/53 [exercises-3-hint-1]
#> 27/53                     
#> 28/53 [unnamed-chunk-5]   
#> 29/53                     
#> 30/53 [exercises-5]       
#> 31/53                     
#> 32/53 [unnamed-chunk-6]   
#> 33/53                     
#> 34/53 [unnamed-chunk-7]   
#> 35/53                     
#> 36/53 [exercises-7]       
#> 37/53                     
#> 38/53 [exercises-8]       
#> 39/53                     
#> 40/53 [exercises-9]       
#> 41/53                     
#> 42/53 [exercises-10]      
#> 43/53                     
#> 44/53 [exercises-11]      
#> 45/53                     
#> 46/53 [exercises-12]      
#> 47/53                     
#> 48/53 [exercises-12-test] 
#> 49/53                     
#> 50/53 [exercises-13]      
#> 51/53                     
#> 52/53 [download-answers]  
#> 
#> 
#> processing file: /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/tutorial.helpers/child_documents/download_answers.Rmd
#> 1/6                   
#> 2/6 [minutes]         
#> 3/6                   
#> 4/6 [unnamed-chunk-38]
#> 5/6                   
#> 6/6 [unnamed-chunk-39]
#> 53/53                     
#> output file: /tmp/RtmpbmX8wm/tutorial.knit.md
#> /opt/hostedtoolcache/pandoc/3.1.11/x64/pandoc +RTS -K512m -RTS /tmp/RtmpbmX8wm/tutorial.knit.md --to html4 --from markdown+autolink_bare_uris+tex_math_single_backslash --output /tmp/RtmpbmX8wm/tutorial.html --lua-filter /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/rmarkdown/rmarkdown/lua/pagebreak.lua --lua-filter /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/rmarkdown/rmarkdown/lua/latex-div.lua --variable bs3=TRUE --standalone --section-divs --template /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/rmarkdown/rmd/h/default.html --no-highlight --variable highlightjs=1 --section-divs --reference-location=section --template /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/learnr/rmarkdown/templates/tutorial/resources/tutorial-format.htm --no-highlight --variable highlightjs=1 --variable progressive=true --variable allow-skip=true --variable learnr-version=0.11.5 --id-prefix section- --variable theme=cerulean --mathjax --variable 'mathjax-url=https://mathjax.rstudio.com/latest/MathJax.js?config=TeX-AMS-MML_HTMLorMML' --include-in-header /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/rmarkdown/rmd/h/shiny-header.html 
#> 
#> Output created: /tmp/RtmpbmX8wm/tutorial.html
#> ── Skip: Rendering /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/tutorial.helpers/tutorials/getting-started/tutorial.Rmd ──
#> Reason: empty test
#> 
#> 
#> 
#> processing file: tutorial.Rmd
#> 1/234                               
#> 2/234 [setup]                       
#> 3/234                               
#> 4/234 [copy-code-chunk]             
#> 
#> 
#> processing file: /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/tutorial.helpers/child_documents/copy_button.Rmd
#> 1/3 [change-code]      
#> 2/3                    
#> 3/3 [observe-transfers]
#> 5/234                               
#> 6/234 [info-section]                
#> 
#> 
#> processing file: /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/tutorial.helpers/child_documents/info_section.Rmd
#> 1/6        
#> 2/6 [name] 
#> 3/6        
#> 4/6 [email]
#> 5/6        
#> 6/6 [ID]   
#> 7/234                               
#> 8/234 [unnamed-chunk-1]             
#> 9/234                               
#> 10/234 [working-with-data-1]         
#> 11/234                               
#> 12/234 [working-with-data-1-hint-1]  
#> 13/234                               
#> 14/234 [working-with-data-1-test]    
#> 15/234                               
#> 16/234 [working-with-data-2]         
#> 17/234                               
#> 18/234 [working-with-data-3]         
#> 19/234                               
#> 20/234 [working-with-data-3-hint-1]  
#> 21/234                               
#> 22/234 [working-with-data-3-test]    
#> 23/234                               
#> 24/234 [working-with-data-4]         
#> 25/234                               
#> 26/234 [working-with-data-4-hint-1]  
#> 27/234                               
#> 28/234 [working-with-data-4-test]    
#> 29/234                               
#> 30/234 [working-with-data-5]         
#> 31/234                               
#> 32/234 [working-with-data-5-hint-1]  
#> 33/234                               
#> 34/234 [working-with-data-5-test]    
#> 35/234                               
#> 36/234 [working-with-data-6]         
#> 37/234                               
#> 38/234 [working-with-data-6-hint-1]  
#> 39/234                               
#> 40/234 [working-with-data-6-test]    
#> 41/234                               
#> 42/234 [working-with-data-7]         
#> 43/234                               
#> 44/234 [working-with-data-7-hint-1]  
#> 45/234                               
#> 46/234 [working-with-data-7-test]    
#> 47/234                               
#> 48/234 [working-with-data-8]         
#> 49/234                               
#> 50/234 [working-with-data-8-hint-1]  
#> 51/234                               
#> 52/234 [working-with-data-8-test]    
#> 53/234                               
#> 54/234 [working-with-data-9]         
#> 55/234                               
#> 56/234 [working-with-data-10]        
#> 57/234                               
#> 58/234 [working-with-data-10-hint-1] 
#> 59/234                               
#> 60/234 [working-with-data-10-test]   
#> 61/234                               
#> 62/234 [working-with-data-11]        
#> 63/234                               
#> 64/234 [working-with-data-12]        
#> 65/234                               
#> 66/234 [scat-p]                      
#> 67/234                               
#> 68/234 [creating-a-graphic-1]        
#> 69/234                               
#> 70/234 [creating-a-graphic-2]        
#> 71/234                               
#> 72/234 [creating-a-graphic-2-hint-1] 
#> 73/234                               
#> 74/234 [creating-a-graphic-2-test]   
#> 75/234                               
#> 76/234 [creating-a-graphic-3]        
#> 77/234                               
#> 78/234 [creating-a-graphic-4]        
#> 79/234                               
#> 80/234 [creating-a-graphic-4-hint-1] 
#> 81/234                               
#> 82/234 [creating-a-graphic-4-test]   
#> 83/234                               
#> 84/234 [creating-a-graphic-5]        
#> 85/234                               
#> 86/234 [creating-a-graphic-5-hint-1] 
#> 87/234                               
#> 88/234 [creating-a-graphic-5-test]   
#> 89/234                               
#> 90/234 [creating-a-graphic-6]        
#> 91/234                               
#> 92/234 [creating-a-graphic-6-hint-1] 
#> 93/234                               
#> 94/234 [creating-a-graphic-6-test]   
#> 95/234                               
#> 96/234 [creating-a-graphic-7]        
#> 97/234                               
#> 98/234 [creating-a-graphic-7-hint-1] 
#> 99/234                               
#> 100/234 [creating-a-graphic-7-test]   
#> 101/234                               
#> 102/234 [creating-a-graphic-8]        
#> 103/234                               
#> 104/234 [creating-a-graphic-8-hint-1] 
#> 105/234                               
#> 106/234 [creating-a-graphic-8-test]   
#> 107/234                               
#> 108/234 [creating-a-graphic-9]        
#> 109/234                               
#> 110/234 [creating-a-graphic-9-hint-1] 
#> 111/234                               
#> 112/234 [creating-a-graphic-9-test]   
#> 113/234                               
#> 114/234 [unnamed-chunk-2]             
#> 115/234                               
#> 116/234 [creating-a-graphic-10]       
#> 117/234                               
#> 118/234 [creating-a-graphic-10-hint-1]
#> 119/234                               
#> 120/234 [creating-a-graphic-10-test]  
#> 121/234                               
#> 122/234 [creating-a-graphic-11]       
#> 123/234                               
#> 124/234 [creating-a-graphic-11-hint-1]
#> 125/234                               
#> 126/234 [creating-a-graphic-11-test]  
#> 127/234                               
#> 128/234 [creating-a-graphic-12]       
#> 129/234                               
#> 130/234 [creating-a-graphic-12-hint-1]
#> 131/234                               
#> 132/234 [creating-a-graphic-12-test]  
#> 133/234                               
#> 134/234 [creating-a-graphic-13]       
#> 135/234                               
#> 136/234 [creating-a-graphic-13-hint-1]
#> 137/234                               
#> 138/234 [creating-a-graphic-13-test]  
#> 139/234                               
#> 140/234 [creating-a-graphic-14]       
#> 141/234                               
#> 142/234 [creating-a-graphic-14-hint-1]
#> 143/234                               
#> 144/234 [creating-a-graphic-14-test]  
#> 145/234                               
#> 146/234 [running-r-code-1]            
#> 147/234                               
#> 148/234 [running-r-code-1-hint-1]     
#> 149/234                               
#> 150/234 [running-r-code-1-test]       
#> 151/234                               
#> 152/234 [running-r-code-2]            
#> 153/234                               
#> 154/234 [running-r-code-2-hint-1]     
#> 155/234                               
#> 156/234 [running-r-code-2-test]       
#> 157/234                               
#> 158/234 [running-r-code-3]            
#> 159/234                               
#> 160/234 [running-r-code-3-hint-1]     
#> 161/234                               
#> 162/234 [running-r-code-3-test]       
#> 163/234                               
#> 164/234 [running-r-code-4]            
#> 165/234                               
#> 166/234 [running-r-code-5]            
#> 167/234                               
#> 168/234 [running-r-code-5-hint-1]     
#> 169/234                               
#> 170/234 [running-r-code-5-test]       
#> 171/234                               
#> 172/234 [running-r-code-6]            
#> 173/234                               
#> 174/234 [running-r-code-6-hint-1]     
#> 175/234                               
#> 176/234 [running-r-code-6-test]       
#> 177/234                               
#> 178/234 [running-r-code-7]            
#> 179/234                               
#> 180/234 [running-r-code-7-hint-1]     
#> 181/234                               
#> 182/234 [running-r-code-7-test]       
#> 183/234                               
#> 184/234 [running-r-code-8]            
#> 185/234                               
#> 186/234 [running-r-code-8-hint-1]     
#> 187/234                               
#> 188/234 [running-r-code-8-test]       
#> 189/234                               
#> 190/234 [running-r-code-9]            
#> 191/234                               
#> 192/234 [running-r-code-9-hint-1]     
#> 193/234                               
#> 194/234 [running-r-code-9-test]       
#> 195/234                               
#> 196/234 [running-r-code-10]           
#> 197/234                               
#> 198/234 [running-r-code-10-hint-1]    
#> 199/234                               
#> 200/234 [running-r-code-10-test]      
#> 201/234                               
#> 202/234 [running-r-code-11]           
#> 203/234                               
#> 204/234 [running-r-code-11-hint-1]    
#> 205/234                               
#> 206/234 [running-r-code-11-test]      
#> 207/234                               
#> 208/234 [running-r-code-12]           
#> 209/234                               
#> 210/234 [running-r-code-12-hint-1]    
#> 211/234                               
#> 212/234 [running-r-code-12-test]      
#> 213/234                               
#> 214/234 [running-r-code-13]           
#> 215/234                               
#> 216/234 [running-r-code-13-hint-1]    
#> 217/234                               
#> 218/234 [running-r-code-13-test]      
#> 219/234                               
#> 220/234 [running-r-code-14]           
#> 221/234                               
#> 222/234 [running-r-code-14-hint-1]    
#> 223/234                               
#> 224/234 [running-r-code-14-test]      
#> 225/234                               
#> 226/234 [unnamed-chunk-3]             
#> 227/234                               
#> 228/234 [running-r-code-15]           
#> 229/234                               
#> 230/234 [running-r-code-15-hint-1]    
#> 231/234                               
#> 232/234 [running-r-code-15-test]      
#> 233/234                               
#> 234/234 [download-answers]            
#> 
#> 
#> processing file: /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/tutorial.helpers/child_documents/download_answers.Rmd
#> 1/6                    
#> 2/6 [minutes]          
#> 3/6                    
#> 4/6 [unnamed-chunk-125]
#> 5/6                    
#> 6/6 [unnamed-chunk-126]
#> output file: /tmp/RtmpbmX8wm/tutorial.knit.md
#> /opt/hostedtoolcache/pandoc/3.1.11/x64/pandoc +RTS -K512m -RTS /tmp/RtmpbmX8wm/tutorial.knit.md --to html4 --from markdown+autolink_bare_uris+tex_math_single_backslash --output /tmp/RtmpbmX8wm/tutorial.html --lua-filter /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/rmarkdown/rmarkdown/lua/pagebreak.lua --lua-filter /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/rmarkdown/rmarkdown/lua/latex-div.lua --variable bs3=TRUE --standalone --section-divs --template /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/rmarkdown/rmd/h/default.html --no-highlight --variable highlightjs=1 --section-divs --reference-location=section --template /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/learnr/rmarkdown/templates/tutorial/resources/tutorial-format.htm --no-highlight --variable highlightjs=1 --variable progressive=true --variable allow-skip=true --variable learnr-version=0.11.5 --id-prefix section- --variable theme=cerulean --mathjax --variable 'mathjax-url=https://mathjax.rstudio.com/latest/MathJax.js?config=TeX-AMS-MML_HTMLorMML' --include-in-header /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/rmarkdown/rmd/h/shiny-header.html 
#> 
#> Output created: /tmp/RtmpbmX8wm/tutorial.html
#> ── Skip: Rendering /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/tutorial.helpers/tutorials/introduction-to-r/tutorial.Rmd ──
#> Reason: empty test
#> 
#> NULL