Knit a set of tutorials

We define "testing" a tutorial as (successfully) running render() on it. This function renders all the tutorials provided in tutorial_paths. There is no check to see if the rendered file looks OK. If a tutorial fails to render, then (we assume!) an error will be generated which will then filter up to our testing rig.

Usage

knit_tutorials(tutorial_paths)

Arguments

tutorial_paths

Character vector of the paths to the tutorials to be knitted.

Value

No return value, called for side effects.

Examples

  knit_tutorials(tutorial_paths = return_tutorial_paths("tutorial.helpers"))
#> 
#> 
#> processing file: tutorial.Rmd
#> 1/53                     
#> 2/53 [setup]             
#> 3/53                     
#> 4/53 [copy-code-chunk]   
#> 
#> 
#> processing file: /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/tutorial.helpers/child_documents/copy_button.Rmd
#> 1/3 [change-code]      
#> 2/3                    
#> 3/3 [observe-transfers]
#> 5/53                     
#> 6/53 [unnamed-chunk-1]   
#> 7/53                     
#> 8/53 [unnamed-chunk-2]   
#> 9/53                     
#> 10/53 [info-section]      
#> 
#> 
#> processing file: /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/tutorial.helpers/child_documents/info_section.Rmd
#> 1/6        
#> 2/6 [name] 
#> 3/6        
#> 4/6 [email]
#> 5/6        
#> 6/6 [ID]   
#> 11/53                     
#> 12/53 [unnamed-chunk-3]   
#> 13/53                     
#> 14/53 [exercises-1]       
#> 15/53                     
#> 16/53 [exercises-1-hint-1]
#> 17/53                     
#> 18/53 [unnamed-chunk-4]   
#> 19/53                     
#> 20/53 [exercises-2]       
#> 21/53                     
#> 22/53 [exercises-2-hint-1]
#> 23/53                     
#> 24/53 [exercises-3]       
#> 25/53                     
#> 26/53 [exercises-3-hint-1]
#> 27/53                     
#> 28/53 [unnamed-chunk-5]   
#> 29/53                     
#> 30/53 [exercises-5]       
#> 31/53                     
#> 32/53 [unnamed-chunk-6]   
#> 33/53                     
#> 34/53 [unnamed-chunk-7]   
#> 35/53                     
#> 36/53 [exercises-7]       
#> 37/53                     
#> 38/53 [exercises-8]       
#> 39/53                     
#> 40/53 [exercises-9]       
#> 41/53                     
#> 42/53 [exercises-10]      
#> 43/53                     
#> 44/53 [exercises-11]      
#> 45/53                     
#> 46/53 [exercises-12]      
#> 47/53                     
#> 48/53 [exercises-12-test] 
#> 49/53                     
#> 50/53 [exercises-13]      
#> 51/53                     
#> 52/53 [download-answers]  
#> 
#> 
#> processing file: /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/tutorial.helpers/child_documents/download_answers.Rmd
#> 1/6                   
#> 2/6 [minutes]         
#> 3/6                   
#> 4/6 [unnamed-chunk-38]
#> 5/6                   
#> 6/6 [unnamed-chunk-39]
#> 53/53                     
#> output file: /tmp/RtmpikMrQV/tutorial.knit.md
#> /opt/hostedtoolcache/pandoc/3.1.11/x64/pandoc +RTS -K512m -RTS /tmp/RtmpikMrQV/tutorial.knit.md --to html4 --from markdown+autolink_bare_uris+tex_math_single_backslash --output /tmp/RtmpikMrQV/tutorial.html --lua-filter /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/rmarkdown/rmarkdown/lua/pagebreak.lua --lua-filter /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/rmarkdown/rmarkdown/lua/latex-div.lua --variable bs3=TRUE --standalone --section-divs --template /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/rmarkdown/rmd/h/default.html --no-highlight --variable highlightjs=1 --section-divs --reference-location=section --template /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/learnr/rmarkdown/templates/tutorial/resources/tutorial-format.htm --no-highlight --variable highlightjs=1 --variable progressive=true --variable allow-skip=true --variable learnr-version=0.11.5 --id-prefix section- --variable theme=cerulean --mathjax --variable 'mathjax-url=https://mathjax.rstudio.com/latest/MathJax.js?config=TeX-AMS-MML_HTMLorMML' --include-in-header /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/rmarkdown/rmd/h/shiny-header.html 
#> 
#> Output created: /tmp/RtmpikMrQV/tutorial.html
#> ── Skip: Rendering /home/runner/.cache/R/renv/library/tutorial.helpers-cdd4ef10/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu/tutorial.helpers/tutorials/getting-started/tutorial.Rmd ──
#> Reason: empty test
#> 
#> NULL